Renamed app folder to src
