+++ app.py
@@ -1,4 +1,4 @@
-from app import create_app
+from src import create_app
 
 if __name__ == "__main__":
     app = create_app("prod")
