+++ src/templates/stock.html
